+++ packages/block-editor/src/components/rich-text/index.js
@@ -10,7 +10,6 @@ import { omit } from 'lodash';
 import { RawHTML, useRef, useCallback, forwardRef } from '@wordpress/element';
 import { useDispatch, useSelect } from '@wordpress/data';
 import {
-	pasteHandler,
 	children as childrenSource,
 	getBlockTransforms,
 	findTransform,
@@ -23,17 +22,13 @@ import {
 	__unstableIsEmptyLine as isEmptyLine,
 	insert,
 	__unstableInsertLineSeparator as insertLineSeparator,
-	create,
-	replace,
 	split,
-	__UNSTABLE_LINE_SEPARATOR as LINE_SEPARATOR,
 	toHTMLString,
 	slice,
 	isCollapsed,
 	removeFormat,
 } from '@wordpress/rich-text';
 import deprecated from '@wordpress/deprecated';
-import { isURL } from '@wordpress/url';
 import { BACKSPACE, DELETE, ENTER } from '@wordpress/keycodes';
 
 /**
@@ -42,18 +37,14 @@ import { BACKSPACE, DELETE, ENTER } from '@wordpress/keycodes';
 import { useBlockEditorAutocompleteProps } from '../autocomplete';
 import { useBlockEditContext } from '../block-edit';
 import { RemoveBrowserShortcuts } from './remove-browser-shortcuts';
-import { filePasteHandler } from './file-paste-handler';
 import FormatToolbarContainer from './format-toolbar-container';
 import { store as blockEditorStore } from '../../store';
 import { useUndoAutomaticChange } from './use-undo-automatic-change';
 import { useCaretInFormat } from './use-caret-in-format';
+import { usePasteHandler } from './use-paste-handler';
 import { useFormatTypes } from './use-format-types';
-import {
-	addActiveFormats,
-	getMultilineTag,
-	getAllowedFormats,
-	isShortcode,
-} from './utils';
+import FormatEdit from './format-edit';
+import { getMultilineTag, getAllowedFormats } from './utils';
 
 function RichTextWrapper(
 	{
@@ -242,126 +233,6 @@ function RichTextWrapper(
 		[ onReplace, onSplit, multilineTag, onSplitMiddle ]
 	);
 
-	const onPaste = useCallback(
-		( {
-			value,
-			onChange,
-			html,
-			plainText,
-			isInternal,
-			files,
-			activeFormats,
-		} ) => {
-			// If the data comes from a rich text instance, we can directly use it
-			// without filtering the data. The filters are only meant for externally
-			// pasted content and remove inline styles.
-			if ( isInternal ) {
-				const pastedValue = create( {
-					html,
-					multilineTag,
-					multilineWrapperTags:
-						multilineTag === 'li' ? [ 'ul', 'ol' ] : undefined,
-					preserveWhiteSpace,
-				} );
-				addActiveFormats( pastedValue, activeFormats );
-				onChange( insert( value, pastedValue ) );
-				return;
-			}
-
-			if ( pastePlainText ) {
-				onChange( insert( value, create( { text: plainText } ) ) );
-				return;
-			}
-
-			// Only process file if no HTML is present.
-			// Note: a pasted file may have the URL as plain text.
-			if ( files && files.length && ! html ) {
-				const content = pasteHandler( {
-					HTML: filePasteHandler( files ),
-					mode: 'BLOCKS',
-					tagName,
-					preserveWhiteSpace,
-				} );
-
-				// Allows us to ask for this information when we get a report.
-				// eslint-disable-next-line no-console
-				window.console.log( 'Received items:\n\n', files );
-
-				if ( onReplace && isEmpty( value ) ) {
-					onReplace( content );
-				} else {
-					splitValue( value, content );
-				}
-
-				return;
-			}
-
-			let mode = onReplace && onSplit ? 'AUTO' : 'INLINE';
-
-			// Force the blocks mode when the user is pasting
-			// on a new line & the content resembles a shortcode.
-			// Otherwise it's going to be detected as inline
-			// and the shortcode won't be replaced.
-			if (
-				mode === 'AUTO' &&
-				isEmpty( value ) &&
-				isShortcode( plainText )
-			) {
-				mode = 'BLOCKS';
-			}
-
-			if (
-				__unstableEmbedURLOnPaste &&
-				isEmpty( value ) &&
-				isURL( plainText.trim() )
-			) {
-				mode = 'BLOCKS';
-			}
-
-			const content = pasteHandler( {
-				HTML: html,
-				plainText,
-				mode,
-				tagName,
-				preserveWhiteSpace,
-			} );
-
-			if ( typeof content === 'string' ) {
-				let valueToInsert = create( { html: content } );
-
-				addActiveFormats( valueToInsert, activeFormats );
-
-				// If the content should be multiline, we should process text
-				// separated by a line break as separate lines.
-				if ( multilineTag ) {
-					valueToInsert = replace(
-						valueToInsert,
-						/\n+/g,
-						LINE_SEPARATOR
-					);
-				}
-
-				onChange( insert( value, valueToInsert ) );
-			} else if ( content.length > 0 ) {
-				if ( onReplace && isEmpty( value ) ) {
-					onReplace( content, content.length - 1, -1 );
-				} else {
-					splitValue( value, content );
-				}
-			}
-		},
-		[
-			tagName,
-			onReplace,
-			onSplit,
-			splitValue,
-			__unstableEmbedURLOnPaste,
-			multilineTag,
-			preserveWhiteSpace,
-			pastePlainText,
-		]
-	);
-
 	const inputRule = useCallback(
 		( value, valueToFormat ) => {
 			if ( ! onReplace ) {
@@ -449,7 +320,6 @@ function RichTextWrapper(
 		onFocus,
 		ref: richTextRef,
 		hasActiveFormats,
-		children: richTextChildren,
 	} = useRichText( {
 		value: adjustedValue,
 		onChange( html, { __unstableFormats, __unstableText } ) {
@@ -462,7 +332,6 @@ function RichTextWrapper(
 		selectionEnd,
 		onSelectionChange,
 		placeholder,
-		onPaste,
 		__unstableIsSelected: isSelected,
 		__unstableInputRule: inputRule,
 		__unstableMultilineTag: multilineTag,
@@ -499,7 +368,8 @@ function RichTextWrapper(
 		if ( event.keyCode === ENTER ) {
 			event.preventDefault();
 
-			const _value = removeEditorOnlyFormats( value );
+			const _value = { ...value };
+			_value.formats = removeEditorOnlyFormats( value );
 			const canSplit = onReplace && onSplit;
 
 			if ( onReplace ) {
@@ -581,7 +451,15 @@ function RichTextWrapper(
 			{ children && children( { value, onChange, onFocus } ) }
 			{ isSelected && <RemoveBrowserShortcuts /> }
 			{ isSelected && autocompleteProps.children }
-			{ isSelected && richTextChildren }
+			{ isSelected && (
+				<FormatEdit
+					value={ value }
+					onChange={ onChange }
+					onFocus={ onFocus }
+					formatTypes={ formatTypes }
+					forwardedRef={ anchorRef }
+				/>
+			) }
 			{ isSelected && hasFormats && (
 				<FormatToolbarContainer
 					inline={ inlineToolbar }
@@ -600,6 +478,21 @@ function RichTextWrapper(
 					props.ref,
 					richTextRef,
 					useUndoAutomaticChange(),
+					usePasteHandler( {
+						isSelected,
+						disableFormats,
+						onChange,
+						value,
+						formatTypes,
+						tagName,
+						onReplace,
+						onSplit,
+						splitValue,
+						__unstableEmbedURLOnPaste,
+						multilineTag,
+						preserveWhiteSpace,
+						pastePlainText,
+					} ),
 					anchorRef,
 					forwardedRef,
 				] ) }
